Code refactoring
Image repository optimization: Now the Pixels repository just make another API call if it's explicit solicited by the constructor or if it has not been done yet, avoiding unnecessary requests and optimizing home loading time when the route changes by the bottomNav.

And other changes
