+++ lib/app/modules/main/modules/home/home_module.dart
@@ -1,15 +1,15 @@
 import 'package:flutter_modular/flutter_modular.dart';
+import 'package:test_modular_app_video_image_api/app/modules/main/repository/exports.dart';
 import 'package:test_modular_app_video_image_api/app/shared/exports.dart';
 
-import 'repository/exports.dart';
 import 'home_controller.dart';
 import 'home_page.dart';
 
 class HomeModule extends Module {
   @override
   List<Bind> get binds => [
-        Bind.factory(
-          (i) => PixelsRepository(),
+        Bind.instance(
+          (i) => i.get<PixelsRepository>(),
         ),
         Bind.singleton(
           (i) => HomeController(
