+++ lib/app/modules/main/repository/media_repository_interface.dart
@@ -5,10 +5,13 @@ abstract class MediaRepositoryI {
     required String search,
     required int perPage,
     required int page,
+    bool update = false,
   });
+
   Future<List<T>> fetchVideos<T>({
     required String search,
     required int perPage,
     required int page,
+    bool update = false,
   });
 }
